First working R substring implementation
The heart of it is a mapping function that
creates a map of the ANSI string, and also a table
that allows conversion of raw and ANSI coordinates.
More development docs are coming soon.

Of course this is slow, so the whole thing must be
rewritten in C.

All tests from the rather extensive crayon substring
test suite pass. \o/